requirements.txt
scapy
google-api-python-client
google-auth-httplib2
google-auth
